+++ src/components/WYSIWYGInput/WYSIWYGInput.stories.tsx
@@ -88,7 +88,12 @@ const HtmlTemplate: Story<WYSIWYGInputProps & { content?: string; className?: st
     <div className="w-full">
       {/* Clear the onHtmlChangeSlow event that storybook provides automatically because it's slow */}
       <WYSIWYGInput ref={inputRef} {...props} onHtmlChangeSlow={undefined} />
-      <Button onClick={() => setHtmlContent(inputRef?.current?.getSanitizedHtml() || '')}>Get HTML</Button>
+      <div className="flex space-x-2 py-2">
+        <Button onClick={() => setHtmlContent(inputRef?.current?.getSanitizedHtml() || '')}>Get HTML</Button>
+        <Button onClick={() => setHtmlContent(inputRef?.current?.resetInitialValue() || '')}>
+          Reset initial value
+        </Button>
+      </div>
       <HTMLParsedContent className="min-w-full max-w-full mt-4" content={htmlContent} />
       <div className="max-w-full mb-4 border-t border-solid border-gray-300">{htmlContent}</div>
     </div>